Stock buy and Sell
Day 7/160: Stock Buy and Sell Problem 💹
Problem Description:
The cost of stock on each day is given in an array prices[]. You are allowed to buy and sell the stock multiple times, but a stock can only be sold if it has been previously bought. The goal is to calculate the maximum profit you can achieve.

Key rules:
1️⃣ You can buy and sell on the same day.
2️⃣ You can only hold one stock at a time.

Example 1:
Input: prices[] = [100, 180, 260, 310, 40, 535, 695]
Output: 865
Explanation:

Buy on day 0 at 100 and sell on day 3 at 310 (Profit: 310 - 100 = 210).
Buy on day 4 at 40 and sell on day 6 at 695 (Profit: 695 - 40 = 655).
Total Profit = 210 + 655 = 865.